Moved general testing code to dedicated file. Adds fairPlayAmsCompatibility property. 401 status support.
